// client/src/typescript-fetch-client/models/AdminControllerListUsersResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Server API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ModelsUser } from './ModelsUser';
import {
    ModelsUserFromJSON,
    ModelsUserFromJSONTyped,
    ModelsUserToJSON,
    ModelsUserToJSONTyped,
} from './ModelsUser';

/**
 * 
 * @export
 * @interface AdminControllerListUsersResponse
 */
export interface AdminControllerListUsersResponse {
    /**
     * 
     * @type {number}
     * @memberof AdminControllerListUsersResponse
     */
    page?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminControllerListUsersResponse
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminControllerListUsersResponse
     */
    totalPages?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminControllerListUsersResponse
     */
    totalRecords?: number;
    /**
     * 
     * @type {Array<ModelsUser>}
     * @memberof AdminControllerListUsersResponse
     */
    users?: Array<ModelsUser>;
}

/**
 * Check if a given object implements the AdminControllerListUsersResponse interface.
 */
export function instanceOfAdminControllerListUsersResponse(value: object): value is AdminControllerListUsersResponse {
    return true;
}

export function AdminControllerListUsersResponseFromJSON(json: any): AdminControllerListUsersResponse {
    return AdminControllerListUsersResponseFromJSONTyped(json, false);
}

export function AdminControllerListUsersResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdminControllerListUsersResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'page': json['page'] == null ? undefined : json['page'],
        'pageSize': json['page_size'] == null ? undefined : json['page_size'],
        'totalPages': json['total_pages'] == null ? undefined : json['total_pages'],
        'totalRecords': json['total_records'] == null ? undefined : json['total_records'],
        'users': json['users'] == null ? undefined : ((json['users'] as Array<any>).map(ModelsUserFromJSON)),
    };
}

export function AdminControllerListUsersResponseToJSON(json: any): AdminControllerListUsersResponse {
    return AdminControllerListUsersResponseToJSONTyped(json, false);
}

export function AdminControllerListUsersResponseToJSONTyped(value?: AdminControllerListUsersResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'page': value['page'],
        'page_size': value['pageSize'],
        'total_pages': value['totalPages'],
        'total_records': value['totalRecords'],
        'users': value['users'] == null ? undefined : ((value['users'] as Array<any>).map(ModelsUserToJSON)),
    };
}