// client/src/typescript-fetch-client/models/ModelsRole.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ModelsRole = {
    RoleUser: 'user',
    RoleAdmin: 'admin',
    RoleDisabled: 'disabled'
} as const;
export type ModelsRole = typeof ModelsRole[keyof typeof ModelsRole];


export function instanceOfModelsRole(value: any): boolean {
    for (const key in ModelsRole) {
        if (Object.prototype.hasOwnProperty.call(ModelsRole, key)) {
            if (ModelsRole[key as keyof typeof ModelsRole] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ModelsRoleFromJSON(json: any): ModelsRole {
    return ModelsRoleFromJSONTyped(json, false);
}

export function ModelsRoleFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelsRole {
    return json as ModelsRole;
}

export function ModelsRoleToJSON(value?: ModelsRole | null): any {
    return value as any;
}

export function ModelsRoleToJSONTyped(value: any, ignoreDiscriminator: boolean): ModelsRole {
    return value as ModelsRole;
}